+++ CHANGELOG.md
@@ -113,9 +113,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.4.0] - 2025-12-01 [RECONNECT]
 
 ### Added
-- Add auto scroll toggle to log panel.
+- Add auto-scroll toggle to log panel.
 - Append a marker line to the log output when an SSH session closes, including a timestamp.
-- Add auto reconnect controls to live logs.
+- Add auto-reconnect controls to live logs.
 - Update log panel frame colors depending if the SSH session is connected (green), re-connecting (yellow) or disconnected (red).
 - Add Clear Logs button.